+++ packages/core/src/generator.js
