+++ packages/core/src/group-info.js
